Started a transition to a more C++ code: Most printfs and files loading/saving were translated to their C++ counterparts (cout, fstream). In most parts this change makes code more readable. The main problem is converting numbers to string. There is one conversion missing when saving states, maybe i will stick with sprintf there.
Some output messages were rewritten to make them more clear.

The input module was modified. Now it uses pollEvents to get keydown and
keyup events from SDL. This makes the save/load-state to work as it
should. The normal GB buttons work as before using GetKeyState.
This new scheme works, but when lot's of keys are pressed (like in a
fighting game) the emulation slows down A LOT. It seems that pollEvent
slows things quite a lot. This should be worked for the next commit.

Code compiles.