Bugs fixed.
Refactoring is really helpfull to find bugs. The little bug that was found in the las commit was just a symptom of a bigger bug. It seems updateModes is not so correct. The problem was trying to draw line 144 which shouldn't be drawn because screen size is 0-143 aka 144 pixels. The bug didn't appear before because all the draw methods made a bounds check. Refactorizing this checks unveiled the bug and now the code is cleaner and in theory this should help performance because a lot less checks are done.

 Also, CmakeList was modified to add the define(LINUX) so pathas are handled correctly on linux.