Sound full refactoring.
Finally cSound was fully refactorized. Now it does not depend on a pointer to mem. Actually, the pointer was removed. All acceses are through read/write byte. Evrey channel now has it's own class.
Channel 4 now uses an Left Feedback Shift Register(LFSR) to generate the pseudo-random wave pattern. The lfsr may be used on the fly, but the array is far more efficient and clear.
Some member functions were moved to the parent cSoundChannel class, some were made pure virtual so it can have a consistent interface. Although each channel is defined directly, this might change in the future.
I made a patchy workaround on Channel1 so fix some problems with the frequency. There is work to do there.