Color Palette bug fixed. HALT instruction fixed.
There was a bug which screwed background and sprites colors. One example was Megaman Extreme. The problem was with registers FF68/69 and FF6A/6B which are used to write to the palette memory. It seems the writes were correct, but reading was plain wrong. It returned the last value written. Now it returns the value the index points at in the palette memory. Megaman Extreme is now fully playable. It seems this affected Pokemon Trading Cards too.

 This fix was found when refactorizing some display registers. There is still a lot of refactorization there, but the path is already set.

 cInput was refactorized to. Now the jpd and jpb global variables were removed and cInput manages it's behavior. Also, cInput was moved to cMemory where it belongs.

 The only test that failed in the Blargg's CPU instructions test was the HALT opcode. Now it finally works correctly. This makes the first full test to actually pass.