Fixed DIV register.
There was a little issue with the DIV register not reseting completely when written. The behavior was modified. Now the Div register is managed by cMemory instead of cCpu. So now cMemory has a new method called updateIO which will manage all the IO update(video, sound, timers and div). The div is now a 16 bit counter which when read, the upper 8 bita are returned. The counter increments the number of cycles executed by the cpu.

 This makes the div_register test from mooneye to pass.