Moved LCDC, LYC, SCX, SCY, WY, WX to display class
* Started moving all display registers to the correct class.

The idea is to move all display logic to its own module (setModes, updateModes, etc). To do this correctly little steps most be carried so nothing breaks.

The move involves setting cases inside memory map and then catching read/write messages on the display class. So the actual state is in the display class and not on IOMap